+++ projects/ngx-lowcode/CHANGELOG.md
@@ -3,3 +3,5 @@
 ## [Unreleased]
 
 - Initial aggregate package release for ngx-lowcode libraries.
+- Bump aggregate package version to `0.1.1` as release batch source.
+- Align release pipeline to sync changed subpackage versions in sandbox and write back published versions via finalize PR.